pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "detr-geo"
version = "0.1.0"
description = "Geospatial object detection using RF-DETR"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Gabriel Price" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: GIS",
]
dependencies = [
    "rasterio>=1.3.0",
    "geopandas>=0.14.0",
    "pyproj>=3.6.0",
    "shapely>=2.0.0",
    "numpy>=1.24.0",
    "Pillow>=9.0.0",
    "tqdm>=4.60.0",
]

[project.urls]
Homepage = "https://github.com/gpriceless/detr-geo"
Repository = "https://github.com/gpriceless/detr-geo"
Issues = "https://github.com/gpriceless/detr-geo/issues"

[project.optional-dependencies]
rfdetr = [
    "rfdetr>=1.4.0,<2.0.0",
    "torch>=2.0.0",
    "supervision>=0.18.0",
]
viz = [
    "leafmap>=0.30.0",
    "matplotlib>=3.7.0",
]
cloud = [
    "pystac>=1.8.0",
]
all = [
    "detr-geo[rfdetr]",
    "detr-geo[viz]",
    "detr-geo[cloud]",
]

[tool.hatch.build.targets.wheel]
packages = ["src/detr_geo"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests that require real test data",
    "gpu: marks tests that require CUDA GPU",
    "slow: marks tests that take more than 10 seconds",
]
addopts = "-ra --strict-markers"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501", "SIM105", "SIM108", "SIM117", "SIM102"]

[tool.ruff.lint.isort]
known-first-party = ["detr_geo"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true
warn_return_any = false
exclude = ["scripts/", "tests/"]
